requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Werkzeug==2.3.7
requests==2.31.0
yfinance==0.2.28
pandas==2.0.3
numpy==1.24.3
feedparser==6.0.10
pytz==2023.3
scikit-learn==1.3.2
tensorflow-cpu==2.13.0
redis==5.0.1
gunicorn==21.2.0